==> R/00_all.qmd <==
---
title: "00_all"
format: 
  html: 
    embed-resources: true
editor: visual
---

```{r, echo = FALSE, warning = FALSE, message = FALSE}
#| output: false
library("quarto")
library("here")
```

```{r, echo = FALSE, warning = FALSE, message = FALSE}
#| output: false

quarto_render(here("R/01_load.qmd"))
```

{{< include 01_load.qmd >}}

```{r, echo = FALSE, warning = FALSE, message = FALSE}
file.rename(here("R/01_load.html"), here("results/01_load.html"))
```

```{r, echo = FALSE, warning = FALSE, message = FALSE}
#| output: false  

quarto_render(here("R/02_clean.qmd"))
```

{{< include 02_clean.qmd >}}

```{r, echo = FALSE, warning = FALSE, message = FALSE}
file.rename(here("R/02_clean.html"), here("results/02_clean.html"))
```

```{r, echo = FALSE, warning = FALSE, message = FALSE}
#| output: false    

quarto_render(here("R/03_augment.qmd"))
```

{{< include 03_augment.qmd >}}

```{r, echo = FALSE, warning = FALSE, message = FALSE}
file.rename(here("R/03_augment.html"), here("results/03_augment.html"))
```

```{r, echo = FALSE, warning = FALSE, message = FALSE}
#| output: false    

quarto_render(here("R/04_describe.qmd"))
```

{{< include 04_describe.qmd >}}

```{r, echo = FALSE, warning = FALSE, message = FALSE}
file.rename(here("R/04_describe.html"), here("results/04_decribe.html"))
```

```{r, echo = FALSE, warning = FALSE, message = FALSE}
#| output: false    

quarto_render(here("R/05_analysis_1.qmd"))
```

{{< include 05_analysis_1.qmd >}}

```{r, echo = FALSE, warning = FALSE, message = FALSE}
file.rename(here("R/05_analysis_1.html"), here("results/05_analysis_1.html"))
```

```{r, echo = FALSE, warning = FALSE, message = FALSE}
#| output: false    

quarto_render(here("R/05_analysis_2.qmd"))
```

{{< include 05_analysis_2.qmd >}}

```{r, echo = FALSE, warning = FALSE, message = FALSE}
file.rename(here("R/05_analysis_2.html"), here("results/05_analysis_2.html"))
```

```{r, echo = FALSE, warning = FALSE, message = FALSE}
#| output: false      

quarto_render(here("R/05_analysis_3.qmd"))
```

{{< include 05_analysis_3.qmd >}}

```{r, echo = FALSE, warning = FALSE, message = FALSE}
file.rename(here("R/05_analysis_3.html"), here("results/05_analysis_3.html"))
```

## 

```{r, echo = FALSE, warning = FALSE, message = FALSE}
#| output: false  
quarto_render(here("R/00_all.qmd"))
```

{{< include 00_all.qmd >}}

```{r}
file.rename(here("R/00_all.html"), here("results/00_all.html"))
```
