==> R/04_describe.qmd <==
---
title: "04_describe"
format: 
  html:
    embed-resources: true
editor: visual
---

## Loading Libraries

The relevant libraries are loaded. When the "tidyverse" library is loaded, the packages included in the tidyverse are displayed. Similarly, when the "here" library is loaded, the directory placement is printed.

```{r}
rm(list=ls())
library("tidyverse")
library("readr")
library("here")
library("scales")
source("99_proj_func.R")
```

## Loading the Long Format, Augmented, Clean Data

```{r, echo = TRUE, warning = FALSE, message = FALSE}

# Loading the clean data tsv file 
cancer_data <- read_tsv(here('data/03_dat_aug.tsv'), show_col_types = FALSE)
```

## Descriptive Statistics of the Data

The Drug data set provides information about various types of medications ranging from treatments of cancer, Alzheimer's disease, and bacterial infections to more simple symptomatic drugs in treating dry cough, common colds, and for pain relief.\
The types of medications vary greatly, therefore this project will only be investigating drugs applicable for the treatment of cancer. This will streamline the project and make it easier to make concrete conclusions.

The purpose of the data is to assist healthcare professionals and patients in making decisions when choosing/using a medication. Therefore, the description and analysis of the data focuses on this perspective and introduces plots that can further aid in this aim.

The data is introduced after both cleaning and augmentation have been performed. The aim of this section is to introduce descriptive statistics in regards to distribution of classes in the different variables.

The raw data set investigates the composition, administration type, applications, and side effects of the different medications. Furthermore, the drugs are reviewed by the user and split into poor, average, and excellent reviews by percentages. The sum of the three categories of reviews adds up to a collective of 100%.

Underneath an elaboration of the specific variables can be found:

-   The active pharmaceutical ingredient (API) describes the main ingredients in a medicine that causes the desired effect. In the cancer data set information can be found about the name, amount, and units of API.

-   The variable Medicine_Name is the trade name of the medication.

-   The variable Manufacturer is the pharmaceutical company that produces the medication.

-   The variable Administration_type explains how the specific medication is taken/used by the patient.

-   The side effect variable is split into number and type, and tells which adverse effects are associated with the medication.

-   Review_level classifies the review of each observation into one of three categories; poor, average, and excellent. Furthermore, Review\_%, describes the percentage of the three classes among each medicine.

-   Classification_Review is a classification variable of review, showing the dominating class of review within each medicine.

-   The variable ID refers to an identifier made in case of medicine names appearing multiple times as unique observations.

### Dimensionality of the Data

Firstly, the dimensionality of the data set is investigated to get an idea of its size.\
The number of variables is shown and a view of the data given.

```{r}
# Checking for dimensionality of dataset 
cancer_data |> 
  dim_desc()

# Showing the data 
cancer_data |> 
  print()

```

The dimensionality of the data set is 155,424 x 20, from which it is given that the data includes 155,424 rows, with 20 variables for each row.

### Number of Unique Cancer Types

It is interesting to look at how many different cancer types are present in the data set and which types it is.

```{r}
# Find unique types of cancer from the Application_type variable
unique_cancer_rows <- cancer_data |>
  # Selecting respective column
  select(Application_type) |>
  # Filtering selected column for rows containing 'cancer'
  filter(str_detect(Application_type, regex("cancer", ignore_case = TRUE))) |>
  # Getting the unique types
  distinct() 

# Number of different cancer types
unique_cancer_rows |>
  dim_desc()

# Overview of the cancer types
unique_cancer_rows |>
  print()

```

From dimensionality of the subset, containing only unique rows with 'cancer' related applications, 23 unique cancer types appears in the cancer data set.

### Distribution of Side Effects

The Count_Side_Effects variable is investigated to find the average number and median of side effects for all the cancer treatment drugs. Furthermore, the maximum and minimum number of side effects is also found.

```{r}
# Summarizing the mean, maximum, and minimum for the Count_Side_Effects variable
SummaryStatistics(cancer_data, Count_Side_Effects)

```

It can be observed that the lowest number of side effects is one and the maximum number of side effects is 23. Furthermore, the average number of side effects for the drugs is around 13. The median of side effects is 10. When comparing the mean and median a slight skewing is observed, as the mean is slightly higher than the median, when also comparing with minimum and maximum values.

### Distribution of Active Ingredients

Now, the API_Count variable is investigated to look into the maximum, minimum, median, and average number of active pharmaceutical ingredients.

```{r}
# Summarizing the mean, maximum, and minimum for the Count_Side_Effects variable
SummaryStatistics(cancer_data, Count_API)

```

The maximum number of active ingredients in the cancer treatment drugs is two and the minimum number is one. The average number of ingredients is close to one, therefore most drugs have one active ingredient. The median and the mean are approximately the same size. When comparing this with the maximum and minimum value a left skewed distribution is indicated, as there otherwise would have been observed an average at approx. 1.5 active pharmaceutical ingredients (APIs).

### Distribution of Application Variables

Lastly, the minimum, maximum, median, and average number of applications are found to see how many different uses the cancer treatment drugs have and how they are distributed.

```{r}
# Summarizing the mean, maximum, and minimum for the Count_Side_Effects variable
SummaryStatistics(cancer_data, Count_Application)

```

There are up to six different applications of the different cancer drugs and minimum one application. The average number of applications for the drugs is around 3, which means most drugs can treat more than one type of cancer. There is a median of 2 applications per drug, slightly lower than the mean, however still somewhat closely related. When comparing with the minimum and maximum value a clear left-skewed distribution appears, as at half of the data lies at 2 applications or less, shown from the median.

Furthermore, the number of unique applications is found to get a picture of the application span of the observations in the data set.

```{r}
# Finding unique drug names from the Medicine_Name variable
unique_application <- cancer_data |>
  filter(!is.na(Application_type)) |>
  select(Application_type) |>
  # Getting the unique types
  distinct() 

# Number of different drugs
unique_application |>
  dim_desc()

# Viewing the different drug names
unique_application |>
  print()
```

There are 49 different applications of the drugs from the cancer data set. Hence, the cancer drugs can also be used for treatment of other diseases than cancer.

### Investigating the Different Drugs

The number of unique drugs in the cancer data set is investigated to get an idea of how many times the different drugs appear. This is shown in a histogram.

```{r}

# Finding unique drug names from the Medicine_Name variable
unique_drugs <- cancer_data |>
  select(Medicine_Name) |>
  # Getting the unique types
  distinct() 

# Number of different drugs
unique_drugs |>
  dim_desc()

# Viewing the different drug names
unique_drugs |>
  print()
```

The dimensions of the cancer data shows that it contains 131 unique drugs related to the application of 23 different types of cancer.

The frequency of each drug is illustrated in a histogram, colored by a threshold of 1,5%. Drugs with a frequency of 1,5% or more are therefore highlighted.

```{r}

# Creating the plot in one pipe
distinct_cancer_data <- cancer_data |>
# Grouping by mecidine name
  distinct(ID,
    .keep_all=TRUE) |> 
  nrow()

Frequency_of_Unique_Drugs_plot <-cancer_data |> 
  distinct(ID,
           Medicine_Name,
           .keep_all=TRUE) |> 
  group_by(Medicine_Name) |> 
# Summarising the % of prevelance of each drug
  summarise(Proportion = n() / distinct_cancer_data) |>
  ungroup() |>
# Making the ggplot of the % prevelance of each drug
  ggplot(aes(x = Medicine_Name, 
             y = Proportion, 
# Coloring by prevelance using threshold of 1.5%
  fill = ifelse(Proportion > 0.015, "orange", "grey"))) +
  geom_bar(stat = "identity", color = "black") + 
  labs(x = "Medicine name",
       y = "Frequency of drug [%]") + 
  theme_minimal() + 
  theme(axis.text.x = element_blank(),
    panel.grid.major.x = element_blank(), 
    panel.grid.minor.x = element_blank()) +
  ggtitle("Frequency of Unique Drugs in Cancer Data") +
  theme(plot.title = element_text(hjust = 0.5)) +
  scale_y_continuous(labels = scales::percent_format()) +
  scale_fill_identity()+
   geom_hline(yintercept = 0.015, linetype = "dashed", color = "darkorange", alpha = 0.7) +
  geom_text(aes(label = ifelse(Proportion > 0.015, as.character(Medicine_Name), "")),
            angle = 90, hjust = 1, size = 4, vjust = -0.5, 
            fontface = "bold", color = "darkorange")

ggsave(here("results/Frequency_of_Unique_Drugs_plot.png"), plot = Frequency_of_Unique_Drugs_plot, width = 6, height = 5)+



print(Frequency_of_Unique_Drugs_plot)

#   
```

The histogram gives an overview of the distribution of the different frequencies for each drug. It is observed that there are 7 drugs that each account for 1,5% or more. Underneath these 7 drugs are extracted, showing the 7 drugs with highest frequency in the data.

```{r}
cancer_data |>
# Grouping by mecidine name
  distinct(ID,
    .keep_all=TRUE) |> 
  # Counting the medications and sort them in descending order
  count(Medicine_Name, sort = TRUE) |>
  # Selecting the first 10 rows 
  head(7)

```

These are the 7 most common drugs for cancer treatment in the data set, based on the frequency.

### Drugs Related to Review Category

The data focus on the relationship between the different drugs and their review percentage in the three categories; poor, average, and excellent. Drugs related to high percentages in each category are therefore found.

#### Drugs with 85% or More Excellent Reviews

```{r}

Best_reviewed_drugs <- cancer_data |>
  filter(Review_level == 'Excellent_Review_%') |>
  filter(`Review_%` >= 85) |>
  select(Medicine_Name, `Review_%`) |>
  distinct()

Best_reviewed_drugs |>
  print()
```

It can be observed that multiple drugs have 100% excellent reviews, meaning these drugs have only gotten good reviews. Overall, there are 11 drugs with 85% or more excellent reviews.

#### Drugs with 85% or More Average Reviews

```{r}

Average_reviewed_drugs <- cancer_data |>
  filter(Review_level == 'Average_Review_%') |>
  filter(`Review_%` >= 85) |>
  select(Medicine_Name, `Review_%`) |>
  distinct()

Average_reviewed_drugs |>
  print()
```

It can be observed that no drugs have 85% or more average reviews, meaning these drugs have gotten more than a total of 15% good and/or bad reviews.

#### Drugs with 85% or More Poor Reviews

```{r}

Worst_reviewed_drugs <- cancer_data |>
  filter(Review_level == 'Poor_Review_%') |>
  filter(`Review_%` >= 85) |>
  select(Medicine_Name, `Review_%`) |>
  distinct()

Worst_reviewed_drugs |>
  print()

```

As with the excellent reviews, the same is true for poor reviews, that some drugs have only poor reviews. There are 7 drugs with 85% or more poor reviews.

## Conclusion

The drug cancer set consists of a wide variety of drugs, that mainly treats cancer, but is also able to treat other diseases. The data set includes both categorical and numeric variables, enabling different kinds of visualizations. The project can now proceed to a more analytic interpretation of the data, to get an even better understanding of the the distribution of the cancer drug test.
