==> R/05_analysis_2.qmd <==
---
title: "05_analysis_2"
format: 
  html:
    embed-resources: true
editor: visual
---

# Analysis of the 7 most prevalent medicines

## Loading library

Loading libraries which will be used for the data analysis. When the "tidyverse" library is loaded, the packages included in the tidyverse are displayed. Similarly, when the "here" library is loaded, the directory placement is printed.

```{r}
rm(list=ls()) 
library("tidyverse") 
library("dplyr") 
library("readr")
library("here")
library("ggridges")
```

## Loading cancer data file

```{r, warning=FALSE, message=FALSE}

# Loading the cleaned and augmented file:  
cancer_data <- read_tsv(here('data/03_dat_aug.tsv'), show_col_types = FALSE) 
```

A subset of the cancer data is made to investigate the 7 most prevalent types of medicine. These medicines were identified through the descriptive statistics and appear to be the only ones exceeding 1.5% prevalence in the cancer data set.

This analysis aims to investigate potential reasons explaining why these exact medicines appear as the most common in treatment of cancer.

### Subset of 7 most prevalent medicines

The subset is made for the 7 most prevalent medicines, by sorting via ID, as some medicine names might appear twice. A list of the medicine names is created, ranked from highest prevalence to lowest. This list is used to subset the cancer data by semi joining the ranked list and the cancer data using ID. Semi join is used in order to join row wise in the data.

```{r}

# Making a total nr. of distinct medicine names from cancer data
distinct_cancer_data <- cancer_data |>
# Grouping by mecidine name
  distinct(ID,
    .keep_all=TRUE) |> 
  nrow()


# Making list of 10 most prevalent drug ID's
Prevalent_10_medicine_names <-cancer_data |> 
  # Sorting by unique drugs
  distinct(ID,
           Medicine_Name,
           .keep_all=TRUE) |> 
  # Searching pr. drug name
  group_by(Medicine_Name) |> 
  # Calculating the % of prevelance of each drug in new variable
  mutate(Proportion = n() / distinct_cancer_data )  |>
  ungroup() |>
  distinct(Medicine_Name, 
           Proportion, 
           .keep_all=TRUE) |>
  filter(Proportion >0.015) |>
  slice_head(n=7)


# Making subset of 10 most prevalent drug IDs
Most_prevalent_medicines <- cancer_data |>
  # Making subset by join using prevalence list
  semi_join(Prevalent_10_medicine_names, 
            by = "ID")

# Viewing our subset of 7 most prevalent drugs
Most_prevalent_medicines |>
  distinct(ID, 
           .keep_all=TRUE) |>
  head(7)
```

The subset shows the 7 most prevalent drugs in the cancer data. From the dimension of the data frame 7 distinct ID's appear

## Relationship between Prevalent Drugs with Corresponding Side Effects and Administration Type

The relationship between the most prevalent medicine names and the number of side effects is investigated, as it is plausible, that a lower number of side effects might explain a higher usage, making the drug more common. As we only aim to investigate unique medicine names, we make sure to distinguish by this in the pipeline of plotting. As the medicines will carry the same API, active pharmaceutical ingredients, the number of side effects will not differ, making the subset valid. The administration type is furthermore highlighted for each drug, as this is also a plausible variable affecting popularity and consumption.

```{r}

# Most_prevalent_medicines_plot <- 
Most_prevalent_medicines |>
  # Sorting and grouping by medicine name 
  distinct(ID,
           .keep_all = TRUE) |>
  # Looking at unique numbers of side effects for each drug
  # Plotting bar plot of side effect number pr. drug
  ggplot(aes(x = Medicine_Name, 
             y = Count_Side_Effects, 
             fill=Administration_type)) + 
  geom_bar(stat= 'identity', 
           alpha = 0.8) + 
  scale_fill_manual(values =c('#A1D99B',
                              '#D6616B',
                              '#414487FF'
                              )) +
  labs(x = "Drug name",
       y = "Number of side effects") + 
  ggtitle(str_wrap("Relationship between side effects and 7 most commmon cancer drugs", 50)) +
  theme_minimal() +
  theme(legend.position = 'bottom') +
  guides(fill=guide_legend(title="Adimnistration type")) +
  theme(panel.grid.major.x = element_blank(),  # Remove major vertical grid lines
        panel.grid.minor.x = element_blank()) 

#ggsave(here("results/Most_prevalent_medicines_plot.png"), plot = Most_prevalent_medicines_plot, width = 6, height = 4)

#print(Most_prevalent_medicines_plot)

```

The visualization shows the 7 different types of drug names, as these unique medicines exceed a prevalence of 1.5%. The distribution of side effects seems slightly random, however with majority having below 10 side effects. It can therefore not directly be concluded that lower number of side effects explains the popularity of the drugs in the data set, but there is a trend of lower side effects for 5/7 drugs.

Furthermore, it appears, from the visualization, that administration types cream, tablet, and injection only appear in the subset. Only one of the drugs is administrated by cream, whereas injection makes up 4 out of 7 drugs, \~55 %, of the administrations whereas tablet makes up 2 incidences, just below \~30%. Hence, popularity of injection is observed, despite being the drug with the most associated side effects.

To investigate this trend further, the density of side effects related to each of the different administration types is visualized in the cancer data with the aim of finding a general trend. This is done for each of the medicines by using the ID in order to obtain unique observations.

```{r}
Side_Effects_Administration_type <- cancer_data |>
   distinct(ID, 
           Administration_type,
           Count_Side_Effects,
           .keep_all = TRUE) |>
  # Using type of administration and side effect number from cancer data
  select(Administration_type, 
         Count_Side_Effects) |>
  # Plotting density of side effects for each administration type 
  ggplot(aes(
    x = Count_Side_Effects, 
    y = as.factor(Administration_type),  
    fill = as.factor(Administration_type)
  )) +
  # Including transparency due to overlap of density plots
  geom_density_ridges(alpha = 0.5) +
  scale_fill_manual(values =c('#20A387FF',
                              '#D6616B',
                              '#A1D99B',
                              '#414487FF', 
                              '#E6550D',
                              '#FDE725FF',
                              '#95D840FF',
                              '#DE9ED6FF' ,
                              '#440154FF' ))  +
  theme_minimal() +
  theme(legend.position = 'none') +
  labs(
    title = "Distribution of Side Effects by Administration Type",
    x = "Number of Side Effects",
    y = "Administration Type",
    fill = "Administration Type"
  ) +
  xlim(0,35)


ggsave(here("results/Side_Effects_Administration_type.png"), plot = Side_Effects_Administration_type, width = 4, height = 4)

print(Side_Effects_Administration_type)
```

The density plot shows a general trend between injection and a medium to high number of side effects. The peak at \~20 confirms what is found in the subset of the prevalent drugs. Furthermore, a peak in the density of side effects around \~8 is observed for tablets, which is in accordance with tablet-administrated drugs Advancan and Rolimus in the prevalence medicine subset.

Lastly, administration through capsules seems to be associated with a high number of side effects whereas administrations such as cream and lotion are associated with a very low number or no side effects.

## Relationship between Prevalent Drugs and Reviews Level

The relationship between the most common cancer drugs and review level is investigated for each unique ID. We look at each ID since we are interested in the prevalence of each review level. The investigation is performed by creating a subset only containing information on ID, medicine name, and classification of review yielding a data frame.

```{r}

Most_prevalent_medicines |>
  # Using ID to dinstinguish observations
  group_by(ID) |> 
  # Selecting variables of interest
  distinct(ID, Medicine_Name, 
           Classification_Review) |>
  # Arranging data in alphabetical order 
  arrange(Classification_Review) |>
  head(10)

```

The data frame shows 7 unique ID's, which can also be found in the descriptive analysis. There is a majority of average reviews for the 7 most prevalent drugs. Whereas only one appears with an excellent review. Two reviews of the category poor are observed as well. This indicates that an excellent reviews does not seem to affect the prescription and/or choice of a drug, however poor might. Furthermore, an average classification seems to be enough for a relatively high distribution. It should be noted that nothing conclusive can be said, merely indicative observations which can be used for further analysis.
