==> R/05_analysis_3.qmd <==
---
title: "05_analysis_3"
format: 
  html:
    embed-resources: true
editor: visual
---

## Analysis of Review Classification

This analysis aims to investigate whether the manufacturer or API has relation to specific review classifications. The top six of each variable are chosen to enable a meaningful and interpretative visualization.

## Loading Libraries and Augmented Data

Firstly, the environment is cleared and libraries and the augmented data set is loaded. When the "tidyverse" library is loaded, the packages included in the tidyverse are displayed. Similarly, when the "here" library is loaded, the directory placement is printed.

```{r, warning=FALSE, message=FALSE}

# Clearing environment
rm(list=ls())

# Loading libraries
library("tidyverse")
library("dplyr")
library("readr")
library("here")

# Loading the augmented data tsv file 
cancer_data <- read_tsv(here('data/03_dat_aug.tsv'))
```

### Influence of Manufacturer

An interesting relation to investigate is if the top six manufacturers have a certain distribution across its reviewed products. The top six manufacturers are those with the most products in the dataset. The distribution of review classifications of the products that each manufacturer produces is shown in a bar plot.

```{r}
# Get the top 6 manufacturers that has the most amount of products in the dataset
top_manufacturers <- cancer_data |> 
  distinct(ID,
           .keep_all = TRUE) |> 
  group_by(Manufacturer) |>
  summarise(product_count = n()) |>   # Count the number of products for each manufacturer
  arrange(desc(product_count)) |>  
  slice_head(n = 6)


# Keeping the rows in cancer data on distinct ID and where the manufactour is one of the top 6 manufacturers
cancer_data_selected <- cancer_data |> 
  distinct(ID,
           .keep_all = TRUE) |> 
  semi_join(top_manufacturers, 
            by = "Manufacturer")

# Reorder the order of Classification_Review factor to ensure "Average" is in the middel.
cancer_data_selected <- cancer_data_selected |> 
  mutate(Classification_Review = factor(Classification_Review, 
                                       levels = c("Poor", 
                                                  "Average", 
                                                  "Excellent")))
# Creating a barplot of the manufactores and their corresponding Review classifications
Manufacturers_bar_plot <- cancer_data_selected |> 
  ggplot(aes(x = Classification_Review, 
             fill = Classification_Review)) +
    geom_bar(stat = "count", 
             show.legend = TRUE) + 
    facet_wrap(~ str_wrap(Manufacturer, 20)) + 
    scale_fill_manual(values = c("Poor" = "#F8766D",  # Making sure color match meaning
                                 "Average" = "#619CFF",
                                 "Excellent" = "#00BA38")) +
    labs(x = '',
         y = "Number of Products", 
         title = str_wrap("Distribution of Review Classification of Products from Top Six  Manufacturers", 55),
         fill= "Review Classification") +
    theme_minimal() +
    theme(axis.text.x=element_blank(),
          legend.position = 'bottom',
          panel.grid.major.x = element_blank(),
      panel.grid.minor.x = element_blank())

# Saving the plot
ggsave(here("results/Manufacturers_bar_plot.png"), plot = Manufacturers_bar_plot, width = 6, height = 4)

print(Manufacturers_bar_plot)
```

From the bar plot it can be seen that there is no clear correlation between being one of the top producers and having a higher number of excellent reviews. Intas Pharmaceuticals Ltd has the most products with poor reviews. Oppositely, Sun Pharmaceuticals Industries Ltd has most products with excellent reviews. Across the distribution of the different reviews there seems to be a trend of there being most Excellent reviews, less Average and Least poor reviews. This is the case for four out of the six top 6 manufactors with most different products.

### Influence of API

In a similar manner the top six used APIs were investigated. Perhaps there is a correlation between being widely used and having good reviews. The bar plot below shows the distribution of reviews for products that includes the top six APIs.

```{r}
# Finging the top 6 most used APIs
top_API <- cancer_data |> 
  filter(!is.na(API_name)) |>   # Exclude rows where API_name is NA
  distinct(ID,
           .keep_all = TRUE) |> 
  group_by(API_name) |> 
  summarise(API_count = n()) |> 
  arrange(desc(API_count)) |>   
  slice_head(n = 6) 

# Keeping the rows in cancer data on distinct ID and where the API_name is one of the top 6 manufacturers
cancer_data_selected <- cancer_data |> 
  distinct(ID,   
           .keep_all = TRUE) |> 
  semi_join(top_API, by = "API_name") |> 
  filter(!is.na(API_name))


# Reorder classification review factor to ensure the average is the middel bar
cancer_data_selected <- cancer_data_selected |> 
  mutate(Classification_Review = factor(Classification_Review, 
                                       levels = c("Poor", 
                                                  "Average", 
                                                  "Excellent")))
# Creating the bar plot
API_bar_plot <-   cancer_data_selected |> 
  ggplot(aes(x = Classification_Review, 
             fill = Classification_Review)) +
    geom_bar(stat = "count", 
             show.legend = TRUE) + 
    facet_wrap(~ str_wrap(API_name, width = 30)) +
    scale_fill_manual(values = c("Poor" = "#F8766D",   # Making sure color match meaning
                                 "Average" = "#619CFF",
                                 "Excellent" = "#00BA38")) +
    labs(x = "", 
         y = "Number of Products", 
         title = "Distribution of Review Classification of Top Six Used APIs",
         fill = "Review Classification") +
    theme_minimal() + 
    theme(axis.text.x=element_blank(),
          legend.position = 'bottom',
          panel.grid.major.x = element_blank(),
      panel.grid.minor.x = element_blank())
          
# Saving the plot          
ggsave(here("results/API_bar_plot.png"), plot = API_bar_plot, width = 6, height = 4)

print(API_bar_plot)
```

It seems there is a random distribution of review classification of the top six most used drugs in the cancer data set. Interestingly, the Recombinant Human Erythropoietin Alfa, which is the top 1 API, has mostly poor and least excellent reviews. Actually, 3 out of 6 APIs has the majority of reviews as poor. Only Leuprolide has excellent as the largest review classifications.

All in all, neither manufacturer or API seems shows a clear tendency in regards to review classification distribution of products from the cancer data set. The conclusion might be different, when looking at the entire data set, while more difficult to visualize and interpret.
